+++ docs/changelog.md
@@ -4,6 +4,8 @@ _This project uses semantic versioning_
 
 ## UNRELEASED
 
+- Upgrade dependencies including [egglog](https://github.com/egraphs-good/egglog/compare/saulshanabrook:egg-smol:a555b2f5e82c684442775cc1a5da94b71930113c...b0db06832264c9b22694bd3de2bdacd55bbe9e32)
+
 ## 8.0.0 (2024-10-17)
 
 - Adds ability to use anonymous functions where callables are needed. These are automatically transformed to egglog